KAN-23: Converted all JS files into TS
