+++ src/pages/UserSignUp.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function UserSignUp() {
+  return <div>UserSignUp</div>;
+}
